Refactor FizzBuzzPortable to improve performance and add extensibility to modulo cases
If a developer suppplied a sufficiently large number (i.e., int.MaxValue) for the upper-bound value, the Execute method would calculate all modulo cases to the large upper-bound value and eventually use so much memory it would trigger an OutOfMemoryException, all without displaying any data to an end user. The usage of the yield keyword within the Execute method resolves this.

Additionally, developers can leverage the FizzBuzzModulo class for creating custom modulo calculations as well as what string values should be returned if the modulo operation is successful.

The WPF demo app now properly renders the result of FizzBuzzPortable.Execute().