Implemented supporting unit tests and refactored classes for clarity
I added a NUnit-based test suite to the solution and implemented unit tests to both test basic functionality and the limits of what these classes could do. As I designed tests, I also found it prudent to rename the classes and touch up docstrings.

While I had intended to use Moq to mock my classes and test to an interface, the classes themselves never reached sufficient complexity for the time limit imposed on the assignment.